Reorganize repo as a lib with example, rather than an umbrella app
In the original configuration, this project was an umbrella app, with
one app being the torch lib itself, and the other app being the example.
This proved problemmatic for a number of reason, so this rather sizeable
commit reorganizes the repo so that the torch lib is at the top level,
and the example app is in the apps dir.