[fix] Update handling of disabled input fields
Previously form `input` and `select` fields were being disabled, regardless of whether the form was being submitted or not.  This means, in case of "validation" failure, the form wasn't submitted, but empty `input` and `select` fields were now disabled and couldn't be edited to bring them into a proper, submittable "valid" state.