The access_no_schema.ex file be mostly blank
Update Torch's copy of the `access_no_schema.ex` file to match what is
in Phoenix by default.  This file is used when no schema is provided, so
it should be a near duplicate of the `schema_access.ex` template.

If there is no schema access, the user would/will need to fill out these
function definitions themselves.