[mix] Update uninstall task to mimic install
We should only uninstall the Torch layout file as
that is all that is created via the corresponding
install task.

The templates are cleaned up each time the `torch.gen.html`
task is run, so if we delete them here from this task,
we may be inadvertently deleting any customizations that the
user has done to the gen.html templates on their own.