Fix ordering of FilterView.find_param for :select
Since the order of the options passed in to the function decide which
value is default, and the "string" input field defaults to "contains",
then the order of the `~w(equals contains)` options passed into
`do_find_param` needs to be switched to `~w(contains equals)`.

Alternatively, if the switch to `~w(equals contains)` was intentional,
the vale passed into `do_find_param` for `:string` should be `equals`
instead of `contains`.